+++ include/openPMD/toolkit/Stdio.hpp
@@ -2,8 +2,6 @@
 
 #include "openPMD/toolkit/ExternalBlockStorage.hpp"
 
-#include <aws/s3/S3Client.h>
-
 namespace openPMD::internal
 {
 struct ExternalBlockStorageStdio : ExternalBlockStorageBackend
@@ -17,18 +15,5 @@ public:
     auto put(std::string const &identifier, void const *data, size_t len)
         -> std::string override;
     ~ExternalBlockStorageStdio() override;
-};
-
-struct ExternalBlockStorageAws : ExternalBlockStorageBackend
-{
-private:
-    Aws::S3::S3Client m_client;
-    std::string m_bucketName;
-
-public:
-    ExternalBlockStorageAws(Aws::S3::S3Client, std::string bucketName);
-    auto put(std::string const &identifier, void const *data, size_t len)
-        -> std::string override;
-    ~ExternalBlockStorageAws() override;
 };
 } // namespace openPMD::internal
